data_migration/university/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0 